refactor: move AbstractQueueTest into a different (non-testing) package refactor: unpdate based on PR comments
