+++ src/main/kotlin/org/wfanet/measurement/gcloud/pubsub/testing/BUILD.bazel
@@ -1,11 +1,11 @@
-load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_test")
+load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_library")
 
 package(
     default_testonly = True,
     default_visibility = ["//visibility:public"],
 )
 
-kt_jvm_test(
+kt_jvm_library(
     name = "google_pub_sub_emulator_client",
     srcs = ["GooglePubSubEmulatorClient.kt"],
     deps = [
@@ -16,7 +16,7 @@ kt_jvm_test(
     ],
 )
 
-kt_jvm_test(
+kt_jvm_library(
     name = "google_pub_sub_emulator_provider",
     srcs = ["GooglePubSubEmulatorProvider.kt"],
     deps = [
