+++ src/main/kotlin/org/wfanet/measurement/queue/testing/BUILD.bazel
@@ -1,15 +1,13 @@
-load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_test")
+load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_library")
 
 package(
     default_testonly = True,
     default_visibility = ["//src/test:__subpackages__"],
 )
 
-kt_jvm_test(
+kt_jvm_library(
     name = "AbstractQueueTest",
     srcs = ["AbstractQueueTest.kt"],
-    local = True,
-    test_class = "org.wfanet.measurement.queue.AbstractQueueTest",
     deps = [
         "//imports/java/com/google/common/truth",
         "//imports/java/org/junit",
