+++ src/test/kotlin/org/wfanet/measurement/gcloud/pubsub/BUILD.bazel
@@ -16,7 +16,7 @@ kt_jvm_test(
         "//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub:subscriber",
         "//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub/testing:google_pub_sub_emulator_client",
         "//src/main/kotlin/org/wfanet/measurement/gcloud/pubsub/testing:google_pub_sub_emulator_provider",
+        "//src/main/kotlin/org/wfanet/measurement/queue/testing:AbstractQueueTest",
         "//src/main/proto/wfa/measurement/queue/testing:test_work_kt_jvm_proto",
-        "//src/test/kotlin/org/wfanet/measurement/queue:AbstractQueueTest",
     ],
 )
